# models/task.go
package models

import (
	"time"

	"github.com/gofrs/uuid"
	"gorm.io/gorm"
)

type Status string

const (
	Pending    Status = "pending"
	InProgress Status = "in_progress"
	Completed  Status = "completed"
)

type Task struct {
	ID          uuid.UUID `json:"id" gorm:"primaryKey"`
	Title       string    `json:"title" gorm:"not null"`
	Description string    `json:"description"`
	Status      Status    `json:"status" gorm:"not null;default:pending"`
	UserID      uuid.UUID `json:"user_id" gorm:"not null;"`
	CreatedAt   time.Time `gorm:"autoCreateTime" json:"created_at"`
	UpdatedAt   time.Time `gorm:"autoUpdateTime" json:"updated_at"`
}

// struct to use when updating a task, since the client should not be able to update
// the ID, UserID, CreatedAt, or UpdatedAt fields.
type TaskUpdate struct {
	Title       *string `json:"title"`
	Description *string `json:"description"`
	Status      *Status `json:"status" binding:"omitempty,oneof=pending in_progress completed"`
}

func (t *Task) BeforeCreate(tx *gorm.DB) (err error) {
	t.ID, err = uuid.NewV4()
	if err != nil {
		return err
	}
	return nil
}